task-94: refactor: добавть uuid для id тостов
