+++ package.json
@@ -57,6 +57,8 @@
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
     "redux": "4.1.1",
+    "redux-thunk": "2.3.0",
+    "uuid": "8.3.2",
     "svg-sprite-loader": "6.0.9",
     "yup": "0.32.9"
   },
@@ -81,8 +83,7 @@
     "@types/react-redux": "7.1.18",
     "@types/react-router-dom": "5.1.8",
     "@types/redux": "3.6.0",
-    "@types/svg-sprite-loader": "3.9.4",
-    "@types/webpack-node-externals": "2.5.2",
+    "@types/uuid": "8.3.1",
     "@typescript-eslint/eslint-plugin": "4.29.0",
     "@typescript-eslint/parser": "4.29.0",
     "autoprefixer": "10.3.1",
