+++ src/AppContext.ts
@@ -4,7 +4,7 @@ import { ToastItem } from '@/components/ui/Toaster/Toast/types';
 
 type AppCtx = Nullable<{
   addToastMessage: (toast: ToastItem) => void;
-  removeToastMessage: (id: number) => void;
+  removeToastMessage: (id: string) => void;
 }>;
 
 const AppContext = createContext({} as AppCtx);
