+++ src/components/ui/Toaster/Toast/Toast.tsx
@@ -1,4 +1,4 @@
-import React, { HTMLAttributes } from 'react';
+import React, { HTMLAttributes, useEffect } from 'react';
 import { block } from 'bem-cn';
 
 import { Button, Icon } from '@/components/ui';
@@ -9,13 +9,27 @@ import closeSvg from 'bootstrap-icons/icons/x-lg.svg';
 
 type ToastProps = {
   toast: ToastItem;
-  deleteToast: (id: number) => void;
+  deleteToast: (id: string) => void;
 } & HTMLAttributes<HTMLDivElement>;
 
 const b = block('toast');
 
 function Toast(props: ToastProps): JSX.Element {
   const { toast, deleteToast } = props;
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      if (toast.timeout) {
+        deleteToast(toast.id);
+      }
+    }, toast.timeout || 1);
+
+    return () => {
+      clearInterval(interval);
+    };
+
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [toast.timeout]);
 
   return (
     <div key={toast.id} className={b()}>
