+++ src/components/ui/Toaster/Toast/types.ts
@@ -1,7 +1,8 @@
 export type ToastItem = {
-  id: number;
+  id: string;
   description: string;
   type: 'success' | 'warning';
+  timeout: number | null;
 };
 
 export type Position = 'top-right' | 'bottom-right' | 'top-left' | 'bottom-left';
