+++ src/components/ui/Toaster/Toaster.tsx
@@ -1,4 +1,4 @@
-import React, { HTMLAttributes, useContext, useEffect } from 'react';
+import React, { HTMLAttributes, useCallback, useContext } from 'react';
 import { block } from 'bem-cn';
 
 import AppContext from '@/AppContext';
@@ -9,32 +9,17 @@ import Toast from './Toast';
 type ToastProps = {
   toastList: Array<ToastItem>;
   position: Position;
-  timeout: number | null;
 } & HTMLAttributes<HTMLDivElement>;
 
 const b = block('toaster');
 
 function Toaster(props: ToastProps): JSX.Element {
-  const { toastList, position, timeout } = props;
+  const { toastList, position } = props;
   const appContext = useContext(AppContext);
 
-  const deleteToast = (id: number) => {
+  const deleteToast = useCallback((id: string) => {
     appContext?.removeToastMessage(id);
-  };
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      if (timeout && toastList.length) {
-        deleteToast(toastList[0].id);
-      }
-    }, timeout || 1);
-
-    return () => {
-      clearInterval(interval);
-    };
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [toastList, timeout]);
+  }, []);
 
   return (
     <div className={b({ position: `${position}` })}>
